// src/app/models/enum/People.ts
export enum PeopleOptions {
    None = 'None',
    LessThan2 = 'LessThan2',
    ThreeToFour = 'ThreeToFour',
    MoreThan4 = 'MoreThan4'
}

export enum TableSize {
    Lessthan2 = "https://localhost:44352/Image/2table.png",
    ThreeToFour = "https://localhost:44352/Image/4table.png",
    MoreThan4 = "https://localhost:44352/Image/8table.png",
    Error = "https://localhost:44352/Image/Error.png"
}